M4/Portafolio/Laboratorio2/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose[cryptography]
